+++ app/src/main/java/com/example/musicdownload/view/fragment/PlayActivity.kt
@@ -1,5 +1,6 @@
 package com.example.musicdownload.view.fragment
 
+import android.app.AlertDialog
 import android.app.Dialog
 import android.content.*
 import android.media.MediaPlayer
@@ -155,7 +156,7 @@ class PlayActivity : AppCompatActivity(), ServiceConnection, MediaPlayer.OnCompl
                     0,
                     false,
                     true,
-                    "favorite",
+                    0,
                     music.name,
                     music.artistName,
                     music.duration,
@@ -516,7 +517,7 @@ class PlayActivity : AppCompatActivity(), ServiceConnection, MediaPlayer.OnCompl
                     0,
                     false,
                     true,
-                    "favorite",
+                    0,
                     music.name,
                     music.artistName,
                     music.duration,
@@ -582,39 +583,49 @@ class PlayActivity : AppCompatActivity(), ServiceConnection, MediaPlayer.OnCompl
         }
 
         viewRemoveDownloadSong?.setOnClickListener {
-            for (i in 0..arrayMusicLocalBase.size - 1) {
-                if (music.name == arrayMusicLocalBase[i].title) {
-                    val fdelete: File = File(arrayMusicLocalBase[i].data)
-                    fdelete.delete()
-                    MediaScannerConnection.scanFile(
-                        this, arrayOf(arrayMusicLocalBase[i].data), null, null
-                    )
-                    getDataStoreEx()
-                    var count2 : Int =0
-                    for (i in 0..listMusicOffline.size-1){
-                        if (music.name == listMusicOffline[i].name){
-                            count2 ++
+            val builder = AlertDialog.Builder(this)
+            builder.setPositiveButton("Yes"){_,_ ->
+                for (i in 0..arrayMusicLocalBase.size - 1) {
+                    if (music.name == arrayMusicLocalBase[i].title) {
+                        val fdelete: File = File(arrayMusicLocalBase[i].data)
+                        fdelete.delete()
+                        MediaScannerConnection.scanFile(
+                            this, arrayOf(arrayMusicLocalBase[i].data), null, null
+                        )
+                        getDataStoreEx()
+                        var count2 : Int =0
+                        for (i in 0..listMusicOffline.size-1){
+                            if (music.name == listMusicOffline[i].name){
+                                count ++
+                            }
                         }
+                        if (count2>0){
+                            viewDownloadSong!!.visibility = View.GONE
+                            tvDownload!!.visibility = View.GONE
+                            imageviewDownloadSong!!.visibility = View.GONE
+                            viewRemoveDownloadSong!!.visibility = View.VISIBLE
+                            removeSong!!.visibility = View.VISIBLE
+                            imageviewRemoveDownloadSong!!.visibility = View.VISIBLE
+                        }
+                        else {
+                            viewDownloadSong!!.visibility = View.VISIBLE
+                            tvDownload!!.visibility = View.VISIBLE
+                            imageviewDownloadSong!!.visibility = View.VISIBLE
+                            viewRemoveDownloadSong!!.visibility = View.GONE
+                            removeSong!!.visibility = View.GONE
+                            imageviewRemoveDownloadSong!!.visibility = View.GONE
+                        }
+                        bottomSheetDialogSong.dismiss()
                     }
-                    if (count2>0){
-                        viewDownloadSong!!.visibility = View.GONE
-                        tvDownload!!.visibility = View.GONE
-                        imageviewDownloadSong!!.visibility = View.GONE
-                        viewRemoveDownloadSong!!.visibility = View.VISIBLE
-                        removeSong!!.visibility = View.VISIBLE
-                        imageviewRemoveDownloadSong!!.visibility = View.VISIBLE
-                    }
-                    else {
-                        viewDownloadSong!!.visibility = View.VISIBLE
-                        tvDownload!!.visibility = View.VISIBLE
-                        imageviewDownloadSong!!.visibility = View.VISIBLE
-                        viewRemoveDownloadSong!!.visibility = View.GONE
-                        removeSong!!.visibility = View.GONE
-                        imageviewRemoveDownloadSong!!.visibility = View.GONE
-                    }
-                    bottomSheetDialogSong.dismiss()
                 }
+                Toast.makeText(this, "Remove song Success",Toast.LENGTH_SHORT).show()
             }
+            builder.setNegativeButton("No"){_,_ ->
+
+            }
+            builder.setTitle("Remove song from this phone?")
+            builder.setMessage("Do you really want??")
+            builder.create().show()
         }
     }
 
@@ -674,7 +685,7 @@ class PlayActivity : AppCompatActivity(), ServiceConnection, MediaPlayer.OnCompl
         if (inputCheck(name)) {
             val playListViewModel = ViewModelProvider(this)[PlayListViewModel::class.java]
             // Create play list
-            val playList = PlayList( name, 0, "")
+            val playList = PlayList(0, name, 0, "")
             // add Data to database
             playListViewModel.addPlayList(playList)
             Toast.makeText(this, "Add PlayList: $name Success", Toast.LENGTH_SHORT).show()
@@ -688,13 +699,14 @@ class PlayActivity : AppCompatActivity(), ServiceConnection, MediaPlayer.OnCompl
     }
 
     private fun insertMusicToPlaylist(music: Music, playList: PlayList) {
+        var count: Int = 0
         val musicPlayListViewModel = ViewModelProvider(this)[MusicPlayListViewModel::class.java]
         //
         val musicPlaylist = MusicPlaylist(
             0,
             false,
             false,
-            playList.name,
+            playList.id,
             music.name,
             music.artistName,
             music.duration,
@@ -704,12 +716,24 @@ class PlayActivity : AppCompatActivity(), ServiceConnection, MediaPlayer.OnCompl
         musicPlayListViewModel.readAllMusicData.observe(this,
             Observer { musicplaylist ->
                 for (item: Int in musicplaylist.indices) {
-                    if (musicplaylist[item].namePlayList == playList.name) {
-                        Toast.makeText(this, "Song ${music.name} already in this Playlist", Toast.LENGTH_SHORT).show()
-                    } else {
-                        musicPlayListViewModel.addMusicPlayList(musicPlaylist)
-                        Toast.makeText(this, "Add PlayList: ${music.name} Success", Toast.LENGTH_SHORT).show()
+                    if (musicplaylist[item].idPlayList == playList.id && musicplaylist[item].name == music.name) {
+                        count++
                     }
+                }
+                Log.e("count", count.toString())
+                if (count > 0) {
+                    Toast.makeText(
+                        this,
+                        "Song ${music.name} already in this Playlist",
+                        Toast.LENGTH_SHORT
+                    ).show()
+                } else {
+                    musicPlayListViewModel.addMusicPlayList(musicPlaylist)
+                    Toast.makeText(
+                        this,
+                        "Add PlayList: ${music.name} Success",
+                        Toast.LENGTH_SHORT
+                    ).show()
                 }
             })
     }
